+++ test/test_tree_1d.jl
@@ -432,6 +432,47 @@ isdir(outdir) && rm(outdir, recursive = true)
         # (e.g., from type instabilities)
         @test_allocations(Trixi.rhs!, semi, sol, 1000)
     end
+
+    @trixi_testset "elixir_shallowwater_rainfall_inclined_plane.jl" begin
+        @test_trixi_include(joinpath(EXAMPLES_DIR,
+                                     "elixir_shallowwater_rainfall_inclined_plane.jl"),
+                            l2=[
+                                0.00032207021140467684,
+                                2.302769240645074e-6,
+                                1.3042254631717584e-15
+                            ],
+                            linf=[
+                                0.0003231150986243637,
+                                2.314345512316177e-6,
+                                4.440892098500626e-15
+                            ],
+                            precipitation_rate=(x, t) -> 1e-3 * t,
+                            tspan=(0.0, 1.0))
+        # Ensure that we do not have excessive memory allocations
+        # (e.g., from type instabilities)
+        @test_allocations(Trixi.rhs!, semi, sol, 1000)
+    end
+
+    @trixi_testset "elixir_shallowwater_rainfall_inclined_plane.jl with GreenAmptModel" begin
+        @test_trixi_include(joinpath(EXAMPLES_DIR,
+                                     "elixir_shallowwater_rainfall_inclined_plane.jl"),
+                            l2=[
+                                6.744115851709234e-5,
+                                3.0877901828495006e-7,
+                                1.3140785893485128e-15
+                            ],
+                            linf=[
+                                6.7565187670141e-5,
+                                8.035004140894328e-7,
+                                4.440892098500626e-15
+                            ],
+                            infiltration_model=GreenAmptModel(3.272e-5, 0.0495, 0.38),
+                            precipitation_rate=(x, t) -> 1e-3 * t,
+                            tspan=(0.0, 1.0))
+        # Ensure that we do not have excessive memory allocations
+        # (e.g., from type instabilities)
+        @test_allocations(Trixi.rhs!, semi, sol, 1000)
+    end
 end # SWE
 
 @testset "Shallow Water Quasi-1D" begin